[refactor] calculations implemented
