+++ src/t-line/DiscoveryModes.ts
@@ -1,0 +1,7 @@
+export enum DiscoveryModes {
+    FOLLOWED_SUBSECTION,
+    RECOMMENDED_SUBSECTION,
+    FOLLOWED_USER,
+    RECOMMENDED_USER,
+    PROMOTED,
+}
